abstraction for cpd with antlr
